Update Intro to R Syntax.Rmd
Hi Cayce!
This file is great! I added a few little things, and a bit of practice with ggplot at the end. Do you think we should include some simple t.tests/ linear regression in this tutorial as well? One option is to just have it as open questions, and if someone doesn;t know how to do it they can stay for the instruction on Friday...